% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoreNewsStat_function.R
\name{MoreNewsStat}
\alias{MoreNewsStat}
\title{MoreNewsStat allows you to extract more stats from more news.}
\usage{
MoreNewsStat(doc_html)
}
\arguments{
\item{doc_html}{input query string}
}
\description{
MoreNewsStat allows you to extract more stats from more news.
}
\examples{
MoreNewsStat()
MoreNewsStat("https://www.theguardian.com/small-business-network/2016/aug/16/how-the-ordnance-survey-is-mapping-out-a-future-for-geotech")
}
\keyword{News}
\keyword{Statistics}
\keyword{Text}

