% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewsStat_function.R
\name{NewsStat}
\alias{NewsStat}
\title{NewsStat}
\usage{
NewsStat(doc_html)
}
\arguments{
\item{doc_html}{input query string}
}
\description{
NewsStat allows you to extract key statistics from a range of online content such as news, websites and blogs.
}
\examples{
NewsStat()
NewsStat("http://www.digitaltrends.com/business/equal-pay-pledge-tech/")
}
\keyword{News}
\keyword{Statistics}
\keyword{Text}

