% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoreNewsStat_function.R
\name{MoreNewsStat}
\alias{MoreNewsStat}
\title{MoreNewsStat is an extension of NewsStat - allowing you to extract more stats from more news.}
\usage{
MoreNewsStat(doc_html)
}
\arguments{
\item{doc_html}{input query string}
}
\description{
MoreNewsStat is an extension of NewsStat - allowing you to extract more stats from more news.
}
\examples{
MoreNewsStat()
MoreNewsStat("https://www.theguardian.com/world/2016/sep/13/new-data-points-to-major-eruption-of-japanese-volcano")
}
\keyword{News}
\keyword{Statistics}
\keyword{Text}

